+++ packages/agent-core-v2/src/app/sessionLegacy/sessionLegacyService.ts
@@ -3,14 +3,15 @@
  *
  * Stateless App-scope dispatcher: each method resolves the target session (and
  * its main agent) per call, delegates to the native v2 services, and projects
- * the result into the v1 wire shape. Only `updateProfile` (the cross-domain
- * `agent_config` patch), `status` (the best-effort status rollup), and `goal`
- * (the current-goal read) live here. No business logic is duplicated here.
+ * the result into the v1 wire shape. Only `status` (the best-effort status
+ * rollup) and `goal` (the current-goal read) live here — the profile route's
+ * title/metadata patch and `agent_config` dispatch are composed by kap-server.
+ * No business logic is duplicated here.
  */
 
 import type { GoalSnapshot } from '#/agent/goal/types';
 
-import type { SessionStatusResponse, UpdateSessionProfileRequest } from './sessionProtocol';
+import type { SessionStatusResponse } from './sessionProtocol';
 import { LifecycleScope } from '#/app/scopes';
 import {
   type IAgentScopeHandle,
@@ -25,7 +26,6 @@ import {
 import { IAgentTokenCountingService } from '#/agent/tokenCounting/tokenCounting';
 import { IAgentGoalService } from '#/agent/goal/goal';
 import { IAgentPermissionModeService } from '#/agent/permissionMode/permissionMode';
-import type { PermissionMode } from '#/agent/permissionPolicy/types';
 import { IAgentPlanService } from '#/features/plan/plan';
 import { IAgentProfileService } from '#/agent/profile/profile';
 import { IAgentSwarmService } from '#/features/swarm/agent/swarm';
@@ -39,10 +39,8 @@ import { ErrorCodes, Error2 } from '#/errors';
 import { ensureMainAgent } from '#/session/agentLifecycle/mainAgent';
 import { IAgentLifecycleService } from '#/session/agentLifecycle/agentLifecycle';
 import { IAgentActivityView } from '#/agent/activityView/activityView';
-import { ISessionContext } from '#/session/sessionContext/sessionContext';
-import { ISessionMetadata } from '#/session/sessionMetadata/sessionMetadata';
 
-import { ISessionLegacyService, type SessionWireFields } from './sessionLegacy';
+import { ISessionLegacyService } from './sessionLegacy';
 
 export class SessionLegacyService implements ISessionLegacyService {
   declare readonly _serviceBrand: undefined;
@@ -57,100 +55,6 @@ export class SessionLegacyService implements ISessionLegacyService {
 
   private resume(sessionId: string): Promise<ISessionScopeHandle | undefined> {
     return resumeSessionById(this.services, sessionId);
-  }
-
-  async updateProfile(
-    sessionId: string,
-    body: UpdateSessionProfileRequest,
-  ): Promise<SessionWireFields> {
-    const session = await this.resume(sessionId);
-    if (session === undefined) {
-      throw new Error2(ErrorCodes.SESSION_NOT_FOUND, `session ${sessionId} does not exist`);
-    }
-    const metadata = session.accessor.get(ISessionMetadata);
-
-    if (typeof body.title === 'string') {
-      await metadata.setTitle(body.title);
-    }
-
-    const metadataPatch = body.metadata;
-    if (metadataPatch !== undefined && Object.keys(metadataPatch).length > 0) {
-      await metadata.update({ custom: { ...(metadataPatch as Record<string, unknown>) } });
-    }
-
-    const agentConfig = body.agent_config;
-    if (agentConfig !== undefined) {
-      const agent = await this.resolveMainAgent(sessionId);
-      await this.applyAgentConfig(agent, agentConfig);
-    }
-
-    const meta = await metadata.read();
-    const ctx = session.accessor.get(ISessionContext);
-    return {
-      id: meta.id,
-      workspaceId: ctx.workspaceId,
-      root: ctx.cwd,
-      title: meta.title,
-      lastPrompt: meta.lastPrompt,
-      createdAt: meta.createdAt,
-      updatedAt: meta.updatedAt,
-      archived: meta.archived,
-      archivedAt: meta.archivedAt,
-      custom: meta.custom,
-    };
-  }
-
-
-  private async applyAgentConfig(
-    agent: IAgentScopeHandle,
-    agentConfig: NonNullable<UpdateSessionProfileRequest['agent_config']>,
-  ): Promise<void> {
-    const profile = agent.accessor.get(IAgentProfileService);
-    if (agentConfig.model !== undefined && agentConfig.model !== '') {
-      await profile.setModel(agentConfig.model);
-    }
-    if (agentConfig.thinking !== undefined) {
-      profile.setThinking(agentConfig.thinking);
-    }
-    if (agentConfig.permission_mode !== undefined) {
-      agent.accessor
-        .get(IAgentLifecycleService)
-        .broadcastPermissionMode(agentConfig.permission_mode as PermissionMode);
-    }
-    if (agentConfig.plan_mode !== undefined) {
-      const plan = agent.accessor.get(IAgentPlanService);
-      const active = (await plan.status()) !== null;
-      if (active !== agentConfig.plan_mode) {
-        if (agentConfig.plan_mode) await plan.enter();
-        else plan.exit();
-      }
-    }
-    if (agentConfig.swarm_mode !== undefined) {
-      const swarm = agent.accessor.get(IAgentSwarmService);
-      if (swarm.isActive !== agentConfig.swarm_mode) {
-        if (agentConfig.swarm_mode) swarm.enter('manual');
-        else swarm.exit();
-      }
-    }
-    if (agentConfig.goal_objective !== undefined) {
-      await agent.accessor
-        .get(IAgentGoalService)
-        .createGoal({ objective: agentConfig.goal_objective });
-    }
-    if (agentConfig.goal_control !== undefined) {
-      const goal = agent.accessor.get(IAgentGoalService);
-      switch (agentConfig.goal_control) {
-        case 'pause':
-          await goal.pauseGoal({});
-          break;
-        case 'resume':
-          await goal.resumeGoal({ continueIfPaused: true, continueIfBlocked: true });
-          break;
-        case 'cancel':
-          await goal.cancelGoal({});
-          break;
-      }
-    }
   }
 
   private async resolveMainAgent(sessionId: string): Promise<IAgentScopeHandle> {
